+++ internal/database/memory/keymanagement/keymanagement.go
@@ -1,4 +1,4 @@
-package memorykms
+package keymanagement
 
 import (
 	"fmt"
